initialize empty arrays to make sure that we always return all fields
If a dataset already has values in one of those list-values (e.g. a relation) and we want to update it via the import that does NOT include any of those values (e.g. no relation): this can only be overwritten, when we make sure to set all the list-values to empty lists before-hand.